Add files via upload
For this project, it will be used the data file TSLA_B to analyse and report all steps required, such as Characterization, Data Preparation as cleaning and renaming, using EDA visualisation, applying PCA, and discussing the Course of Dimensionality.
First of all, it is essential to know what is TESLA and what this dataset is showing us. About TESLA, was founded in 2003. It is a North American automotive and energy storage company. Today is accelerating efforts to sustainable energy, builds not only all-electric vehicles but also is investing to produce infinitely scalable clean energy generation and storage products. (About Tesla | Tesla, 2021)